pyproject.toml
[project]
name = "llms-txt-mcp"
version = "0.1.0"
description = "Lean MCP server for minimal-context docs via llms.txt"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Misha Kolesnik", email = "misha@kolesnik.io"}
]
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.12.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "PyYAML>=6.0.0",
    "sentence-transformers>=3.0.0",
    "chromadb>=0.5.0",
    "llms-txt>=0.0.4",
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Build Tools",
]
keywords = [
    "mcp",
    "model context protocol",
    "claude code",
    "documentation",
    "llms.txt",
    "semantic search",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
]

[project.scripts]
llms-txt-mcp = "llms_txt_mcp.server:main"

[project.urls]
Homepage = "https://github.com/tenequm/llms-mcp-txt"
Repository = "https://github.com/tenequm/llms-mcp-txt"
Issues = "https://github.com/tenequm/llms-mcp-txt/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llms_txt_mcp"]

[tool.hatch.build.targets.sdist]
exclude = [
    "tests/**",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
asyncio_mode = "auto"
addopts = "--tb=short --strict-markers"
markers = [
    "performance: marks tests as performance tests (may be slower)",
    "format_detection: tests for detecting llms.txt formats",
    "parsing: tests for parsing llms.txt content",
    "mcp_tools: tests for MCP tool functionality",
    "integration: integration tests"
]

